# specs/002-yaml-catalog/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QAUserSearch Catalog API
  description: API para consulta de catálogo de metadados de fontes externas
  version: 2.0.0
  contact:
    name: QA Team

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /api/v1/catalog/sources:
    get:
      summary: Listar sources do catálogo
      description: Retorna lista paginada de todas as sources catalogadas
      operationId: listSources
      tags:
        - catalog
      parameters:
        - name: db_name
          in: query
          description: Filtrar por nome do banco de dados
          required: false
          schema:
            type: string
        - name: skip
          in: query
          description: Número de registros a pular
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Número máximo de registros a retornar
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: Lista de sources
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceListResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/catalog/sources/{source_id}:
    get:
      summary: Obter detalhes de uma source
      description: Retorna informações detalhadas de uma source incluindo colunas
      operationId: getSource
      tags:
        - catalog
      parameters:
        - name: source_id
          in: path
          description: ID da source (formato db_name.table_name)
          required: true
          schema:
            type: string
            example: "credit.invoice"
      responses:
        '200':
          description: Detalhes da source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceDetailResponse'
        '404':
          description: Source não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/catalog/sources/{source_id}/columns:
    get:
      summary: Listar colunas de uma source
      description: Retorna lista paginada de colunas com filtros opcionais
      operationId: listColumns
      tags:
        - catalog
      parameters:
        - name: source_id
          in: path
          description: ID da source (formato db_name.table_name)
          required: true
          schema:
            type: string
            example: "credit.invoice"
        - name: type
          in: query
          description: Filtrar por tipo inferido
          required: false
          schema:
            $ref: '#/components/schemas/InferredType'
        - name: is_required
          in: query
          description: Filtrar por campos obrigatórios
          required: false
          schema:
            type: boolean
        - name: is_enumerable
          in: query
          description: Filtrar por campos enumeráveis
          required: false
          schema:
            type: boolean
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: Lista de colunas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnListResponse'
        '404':
          description: Source não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    InferredType:
      type: string
      enum:
        - string
        - integer
        - number
        - boolean
        - datetime
        - objectid
        - array
        - object
        - "null"
        - unknown

    EnrichmentStatus:
      type: string
      enum:
        - not_enriched
        - pending_enrichment
        - enriched

    SourceSummary:
      type: object
      required:
        - id
        - db_name
        - table_name
        - column_count
        - enumerable_count
        - cataloged_at
        - updated_at
      properties:
        id:
          type: string
          description: ID da source (formato db_name.table_name)
          example: "credit.invoice"
        db_name:
          type: string
          example: "credit"
        table_name:
          type: string
          example: "invoice"
        column_count:
          type: integer
          minimum: 0
        enumerable_count:
          type: integer
          minimum: 0
        cataloged_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    SourceListResponse:
      type: object
      required:
        - items
        - total
        - skip
        - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SourceSummary'
        total:
          type: integer
          minimum: 0
        skip:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1

    ColumnDetail:
      type: object
      required:
        - id
        - column_name
        - column_path
        - inferred_type
        - is_required
        - is_nullable
        - is_enumerable
        - presence_ratio
        - enrichment_status
      properties:
        id:
          type: string
          description: ID único da coluna (source_id.column_path)
          example: "credit.invoice._id"
        column_name:
          type: string
        column_path:
          type: string
        inferred_type:
          $ref: '#/components/schemas/InferredType'
        is_required:
          type: boolean
        is_nullable:
          type: boolean
        is_enumerable:
          type: boolean
        unique_values:
          type: array
          items: {}
          nullable: true
        sample_values:
          type: array
          items: {}
          nullable: true
        presence_ratio:
          type: number
          minimum: 0
          maximum: 1
        description:
          type: string
          nullable: true
        enrichment_status:
          $ref: '#/components/schemas/EnrichmentStatus'

    ColumnListResponse:
      type: object
      required:
        - items
        - total
        - skip
        - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ColumnDetail'
        total:
          type: integer
          minimum: 0
        skip:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1

    SourceStats:
      type: object
      required:
        - total_columns
        - required_columns
        - enumerable_columns
        - types_distribution
      properties:
        total_columns:
          type: integer
          minimum: 0
        required_columns:
          type: integer
          minimum: 0
        enumerable_columns:
          type: integer
          minimum: 0
        types_distribution:
          type: object
          additionalProperties:
            type: integer

    SourceDetailResponse:
      type: object
      required:
        - id
        - db_name
        - table_name
        - document_count
        - cataloged_at
        - updated_at
        - columns
        - stats
      properties:
        id:
          type: string
          example: "credit.invoice"
        db_name:
          type: string
        table_name:
          type: string
        document_count:
          type: integer
          minimum: 0
        cataloged_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: '#/components/schemas/ColumnDetail'
        stats:
          $ref: '#/components/schemas/SourceStats'

    ErrorResponse:
      type: object
      required:
        - detail
        - error_code
        - timestamp
      properties:
        detail:
          type: string
        error_code:
          type: string
        timestamp:
          type: string
          format: date-time

tags:
  - name: catalog
    description: Operações de catálogo de metadados